// client/src/api/models/postComment.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Readdit
 * OpenAPI spec version: v1
 */
import type { CommunityPost } from "./communityPost";
import type { ApplicationUser } from "./applicationUser";
import type { CommentVote } from "./commentVote";

export interface PostComment {
   id?: string | null;
   createdOn?: string;
   modifiedOn?: string | null;
   isDeleted?: boolean;
   deletedOn?: string | null;
   postId: string;
   post?: CommunityPost;
   authorId: string;
   author?: ApplicationUser;
   content: string;
   voteScore: number;
   votes?: CommentVote[] | null;
}